requirements.txt
pyyaml
requests
python-dotenv
atlassian-python-api
fastapi
uvicorn
pydantic

